added delete forum to admin panel, fixed user search in admin panel, fixed delete comment, delete post

+++ src/server/delete-comment.php
@@ -9,10 +9,20 @@ require('db_connect.php');
 if ($error) {
   exit(json_encode(['status' => 'error', 'message' => $error]));
 }
+// Get post_id of comment
+$result = mysqli_query($connection, "SELECT post_id FROM comments WHERE id = $_POST[comment_id]");
+if (!$result) {
+  exit(json_encode(['status' => 'error', 'message' => mysqli_error($conn)]));
+}
+$post = mysqli_fetch_assoc($result)['post_id'];
 $comment_id = $_POST['comment_id'];
 $result = mysqli_query($connection, "DELETE FROM comments WHERE id = $comment_id");
 if (!$result) {
   exit(json_encode(['status' => 'error', 'message' => 'Error deleting comment']));
 }
+$result = mysqli_query(
+  $connection,
+  "UPDATE posts SET comment_count = comment_count - 1 WHERE id = $post"
+);
 mysqli_close($connection);
 exit(json_encode(['status' => 'success', 'message' => 'Comment deleted']));
